requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncio==3.4.3
attrs==22.1.0
chardet==4.0.0
discord.py==1.7.3
idna==3.3
multidict==6.0.2
numpy==1.23.1
pandas==1.4.3
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
typing_extensions==4.3.0
yarl==1.8.1
f
